legal_agent_final/requirements.txt
flask==3.0.3
flask-cors==4.0.1
PyPDF2==3.0.1
langchain==0.2.14
langchain-core==0.2.38
langchain-community==0.2.12
langchain-google-vertexai==1.0.6
google-cloud-aiplatform==1.71.1
google-auth==2.35.0
pydantic>=2.7,<3.0
grpcio==1.62.2
zstandard==0.22.0
google-cloud-vision==3.8.1
google-cloud-documentai==2.29.0
